chore: seperate withAssessments from withAssessment
