+++ src/components/with-assessment/redux/action-types.ts
@@ -4,9 +4,3 @@ export const GET_ASSESSMENT_FAILED = 'GET_ASSESSMENT_FAILED' as const;
 export const GET_CATALOG_RATING_REQUESTED = 'GET_CATALOG_RATING_REQUESTED' as const;
 export const GET_CATALOG_RATING_SUCCEEDED = 'GET_CATALOG_RATING_SUCCEEDED' as const;
 export const GET_CATALOG_RATING_FAILED = 'GET_CATALOG_RATING_FAILED' as const;
-export const GET_ASSESSMENTS_REQUESTED = 'GET_ASSESSMENTS_REQUESTED' as const;
-export const GET_ASSESSMENTS_SUCCEEDED = 'GET_ASSESSMENTS_SUCCEEDED' as const;
-export const GET_ASSESSMENTS_FAILED = 'GET_ASSESSMENTS_FAILED' as const;
-export const LOAD_MORE_ASSESSMENTS_REQUESTED = 'LOAD_MORE_ASSESSMENTS_REQUESTED' as const;
-export const LOAD_MORE_ASSESSMENTS_SUCCEEDED = 'LOAD_MORE_ASSESSMENTS_SUCCEEDED' as const;
-export const LOAD_MORE_ASSESSMENTS_FAILED = 'LOAD_MORE_ASSESSMENTS_FAILED' as const;
