+++ src/components/with-assessment/redux/actions.ts
@@ -2,18 +2,12 @@ import {
   GET_ASSESSMENT_REQUESTED,
   GET_ASSESSMENT_SUCCEEDED,
   GET_ASSESSMENT_FAILED,
-  GET_ASSESSMENTS_REQUESTED,
-  GET_ASSESSMENTS_SUCCEEDED,
   GET_CATALOG_RATING_REQUESTED,
   GET_CATALOG_RATING_SUCCEEDED,
-  LOAD_MORE_ASSESSMENTS_REQUESTED,
-  LOAD_MORE_ASSESSMENTS_SUCCEEDED,
-  LOAD_MORE_ASSESSMENTS_FAILED,
-  GET_ASSESSMENTS_FAILED,
   GET_CATALOG_RATING_FAILED
 } from './action-types';
 
-import type { Assessment, PagedAssessments, Rating } from '../../../types';
+import type { Assessment, Rating } from '../../../types';
 
 export function getAssessmentRequested(id: string) {
   return {
@@ -67,90 +61,6 @@ export function getCatalogRatingSucceeded(catalogRating: Rating) {
 export function getCatalogRatingFailed(message: string) {
   return {
     type: GET_CATALOG_RATING_FAILED,
-    payload: {
-      message
-    }
-  };
-}
-
-export function getPagedAssessmentsRequested(
-  catalogId: string,
-  entityType: string,
-  context: string,
-  page: number
-) {
-  return {
-    type: GET_ASSESSMENTS_REQUESTED,
-    payload: {
-      catalogId,
-      entityType,
-      context,
-      page
-    }
-  };
-}
-
-export function getPagedAssessmentsSucceeded(
-  pagedAssessments: PagedAssessments
-) {
-  return {
-    type: GET_ASSESSMENTS_SUCCEEDED,
-    payload: {
-      assessments: pagedAssessments.content,
-      totalAssessments: pagedAssessments.totalElements,
-      assessmentPageSize: pagedAssessments.size,
-      assessmentsPage: pagedAssessments.number,
-      hasMoreAssessments:
-        pagedAssessments.number + 1 < pagedAssessments.totalPages
-    }
-  };
-}
-
-export function getPagedAssessmentsFailed(message: string) {
-  return {
-    type: GET_ASSESSMENTS_FAILED,
-    payload: {
-      message
-    }
-  };
-}
-
-export function loadMoreAssessmentsRequested(
-  catalogId: string,
-  entityType: string,
-  context: string,
-  page: number
-) {
-  return {
-    type: LOAD_MORE_ASSESSMENTS_REQUESTED,
-    payload: {
-      catalogId,
-      entityType,
-      context,
-      page
-    }
-  };
-}
-
-export function loadMoreAssessmentsSucceeded(
-  pagedAssessments: PagedAssessments
-) {
-  return {
-    type: LOAD_MORE_ASSESSMENTS_SUCCEEDED,
-    payload: {
-      assessments: pagedAssessments.content,
-      totalAssessments: pagedAssessments.totalElements,
-      assessmentPageSize: pagedAssessments.size,
-      assessmentsPage: pagedAssessments.number,
-      hasMoreAssessments:
-        pagedAssessments.number + 1 < pagedAssessments.totalPages
-    }
-  };
-}
-
-export function loadMoreAssessmentsFailed(message: string) {
-  return {
-    type: LOAD_MORE_ASSESSMENTS_FAILED,
     payload: {
       message
     }
