+++ src/components/with-assessment/redux/saga.ts
@@ -5,13 +5,11 @@ import { getConfig } from '../../../config';
 
 import {
   GET_ASSESSMENT_REQUESTED,
-  GET_ASSESSMENTS_REQUESTED,
-  GET_CATALOG_RATING_REQUESTED,
-  LOAD_MORE_ASSESSMENTS_REQUESTED
+  GET_CATALOG_RATING_REQUESTED
 } from './action-types';
 import * as actions from './actions';
 
-import type { Assessment, PagedAssessments, Rating } from '../../../types';
+import type { Assessment, Rating } from '../../../types';
 
 function* getAssessmentRequested({
   payload: { id }
@@ -60,65 +58,9 @@ function* getCatalogRatingRequested({
   }
 }
 
-function* getPagedAssessmentsRequested({
-  payload: { catalogId, entityType, context, page }
-}: ReturnType<typeof actions.getPagedAssessmentsRequested>) {
-  try {
-    const { data } = yield call(
-      axios.get,
-      `${getConfig().metadataQualityAssessmentApi.host}/assessments/entities`,
-      {
-        params: {
-          catalogId,
-          entityType,
-          context,
-          page
-        }
-      }
-    );
-
-    if (data) {
-      yield put(actions.getPagedAssessmentsSucceeded(data as PagedAssessments));
-    } else {
-      yield put(actions.getPagedAssessmentsFailed(''));
-    }
-  } catch (e) {
-    yield put(actions.getPagedAssessmentsFailed(e.message));
-  }
-}
-
-function* loadMoreAssessmentsRequested({
-  payload: { catalogId, entityType, context, page }
-}: ReturnType<typeof actions.loadMoreAssessmentsRequested>) {
-  try {
-    const { data } = yield call(
-      axios.get,
-      `${getConfig().metadataQualityAssessmentApi.host}/assessments/entities`,
-      {
-        params: {
-          catalogId,
-          entityType,
-          context,
-          page
-        }
-      }
-    );
-
-    if (data) {
-      yield put(actions.loadMoreAssessmentsSucceeded(data as PagedAssessments));
-    } else {
-      yield put(actions.loadMoreAssessmentsFailed(''));
-    }
-  } catch (e) {
-    yield put(actions.loadMoreAssessmentsFailed(e.message));
-  }
-}
-
 export default function* saga() {
   yield all([
     takeLatest(GET_ASSESSMENT_REQUESTED, getAssessmentRequested),
-    takeLatest(GET_CATALOG_RATING_REQUESTED, getCatalogRatingRequested),
-    takeLatest(GET_ASSESSMENTS_REQUESTED, getPagedAssessmentsRequested),
-    takeLatest(LOAD_MORE_ASSESSMENTS_REQUESTED, loadMoreAssessmentsRequested)
+    takeLatest(GET_CATALOG_RATING_REQUESTED, getCatalogRatingRequested)
   ]);
 }
