+++ src/components/with-assessments/redux/action-types.ts
@@ -1,0 +1,6 @@
+export const GET_ASSESSMENTS_REQUESTED = 'GET_ASSESSMENTS_REQUESTED' as const;
+export const GET_ASSESSMENTS_SUCCEEDED = 'GET_ASSESSMENTS_SUCCEEDED' as const;
+export const GET_ASSESSMENTS_FAILED = 'GET_ASSESSMENTS_FAILED' as const;
+export const LOAD_MORE_ASSESSMENTS_REQUESTED = 'LOAD_MORE_ASSESSMENTS_REQUESTED' as const;
+export const LOAD_MORE_ASSESSMENTS_SUCCEEDED = 'LOAD_MORE_ASSESSMENTS_SUCCEEDED' as const;
+export const LOAD_MORE_ASSESSMENTS_FAILED = 'LOAD_MORE_ASSESSMENTS_FAILED' as const;
