+++ src/redux/saga.ts
@@ -12,6 +12,7 @@ import organizationsSaga from '../components/with-organizations/redux/saga';
 import organizationSaga from '../components/with-organization/redux/saga';
 import reportSaga from '../components/with-report/redux/saga';
 import assessmentSaga from '../components/with-assessment/redux/saga';
+import assessmentsSaga from '../components/with-assessments/redux/saga';
 import informationModelSaga from '../components/with-information-model/redux/saga';
 import informationModelsSaga from '../components/with-information-models/redux/saga';
 import organizationsCatalogSaga from '../components/with-organizations-catalog/redux/saga';
@@ -37,6 +38,7 @@ export default function* saga() {
     organizationSaga(),
     reportSaga(),
     assessmentSaga(),
+    assessmentsSaga(),
     informationModelSaga(),
     informationModelsSaga(),
     organizationsCatalogSaga(),
